+++ SampleProject/StartUpViewController.swift
@@ -25,21 +25,7 @@ class StartUpViewController: UIViewController {
         .disposed(by: disposeBag)
     }
     private func login() {
-        let baseUrl = "http://zipcloud.ibsnet.co.jp/api/"
-        let searchUrl = "\(baseUrl)search"
-        let parameters: [String: Any] = ["zipcode": "2790031"]
-        let headers: HTTPHeaders = ["Content-Type": "application/json"]
-        AF.request(searchUrl, method: .get, parameters: parameters, encoding: URLEncoding(destination: .queryString), headers: headers).responseJSON { response in
-            guard let data = response.data else {
-                return
-            }
-            print(response)
-//            do {
-//                self.addresses = try JSONDecoder().decode(AddressModel.self, from: data)
-//            } catch let error {
-//                print("Error: \(error)")
-//            }
-        }
+        ApiProvider.requestApi()
     }
     /*
     // MARK: - Navigation
